requirements.txt
pyusb>=1.2.1
PyQt6>=6.6.0
